chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
python-multipart
passlib[bcrypt]
fastapi-jwt-auth
pydantic<2.0
sentence-transformers
faiss-cpu
scikit-learn
joblib
Pillow
pytesseract
PyMuPDF
azure-storage-blob
openai
atlassian-python-api
requests
